model2/CNNmodels/service/streamlit_requirements.txt
fastapi==0.82.0
uvicorn==0.18.3
mlfoundry>=0.4.3,<0.5
pandas==1.4.4
numpy==1.23.3
tensorflow==2.9.1
opencv-python==4.6.0.66
tensorflow-addons==0.17.1
adabelief-tf==0.2.1
streamlit
pillow
matplotlib